schema/pyproject.toml
[tool.poetry]
name = "owasp-schemas"
version = "0.1.0"
description = "A collection of OWASP schemas"
authors = ["Arkadii Yakovets <arkadii.yakovets@owasp.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
jsonschema = "^4.23.0"
pytest = "^8.3.4"
python = "^3.13"
pyyaml = "^6.0.2"
validators = "^0.34.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
